+++ packages/apps/crm/src/components/organisms/index.js
@@ -16,3 +16,4 @@ export {default as ClientBottom} from './ClientBottom/ClientBottom';
 export {default as ContactHeader} from './ContactHeader/ContactHeader';
 export {default as ContactBody} from './ContactBody/ContactBody';
 export {default as ContactBottom} from './ContactBottom/ContactBottom';
+export {default as EventHeader} from './EventHeader/EventHeader';
